docs(stack-control): 024 phase 9 — compass surface + honest-boundary docs; line-cap (US, FR-014)
- README: new "Lifecycle compass — the un-skippable workflow" section documenting the
  `workflow compass` verb (orient + intent diff + gating exit codes), the fixed intent
  vocabulary, the per-skill hard-refusal embedding, the back-half gate, and the
  orphan-is-off-rail rule. Records the FR-014 honest boundary explicitly: the
  mechanism binds an agent following its skills, NOT a human with raw git/gh or a
  direct /speckit-implement — not overclaimed.
- Line-cap (finding C1): moved resolveConvergenceItem out of the over-cap govern.ts
  into feature-resolution.ts (its proper home); govern.ts 1019 → 994. incremental-audit
  stayed at 202 (FR-012 fix did not grow it past cap); payload-implement untouched.

Verified: full suite 1681 green; real-installation dogfood of `workflow compass`
(orientation, intent on-course, off-rail exit 4, unknown-intent exit 2) on the
session-pinned branch. Zero new tsc errors vs HEAD.
